requirements.txt
protobuf
# Core dependencies
streamlit>=1.50.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.8.0

# LLM providers
openai>=1.109.0
google-generativeai>=0.8.5

# Data handling
pydantic>=2.0.0

# Audio/TTS dependencies
streamlit-webrtc>=0.47.0
aiortc>=1.7.0
av>=12.0.0
numpy>=1.24.0
elevenlabs>=1.0.0
websockets>=12.0